add dto, refactor services

+++ src/Controller/UserController.php
@@ -3,15 +3,19 @@
 namespace App\Controller;
 
 use App\Service\User\UserManagementService;
+use Src\Dto\UserManagementDto;
+use Src\Exception\UserManagementException;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\Validator\Validator\ValidatorInterface;
 
 final class UserController extends AbstractController
 {
     public function __construct(
-        private UserManagementService $managementService
+        private UserManagementService $managementService,
+        private ValidatorInterface $validator
     ) {
     }
 
@@ -23,52 +27,75 @@ final class UserController extends AbstractController
         ]);
     }
 
+    private function hasErrors(UserManagementDto $dto)
+    {
+        $errors = $this->validator->validate($dto);
+
+        if (\count($errors) === 0)
+            return false;
+
+        return true;
+    }
+
+    private function handleRequest(Request $request, string $action)
+    {
+        $dto = UserManagementDto::fromRequest($request);
+
+        if ($this->hasErrors($dto))
+            return $this->json(['status' => 'error'], Response::HTTP_BAD_REQUEST);
+
+        try {
+
+            match ($action) {
+                'block' =>
+                $this->managementService->blockUser(
+                    $dto->userIds
+                ),
+
+                'unblock' =>
+                $this->managementService->unblockUser(
+                    $dto->userIds
+                ),
+
+                'delete' =>
+                $this->managementService->deleteUser(
+                    $dto->userIds
+                )
+            };
+
+        } catch (UserManagementException $e) {
+            return $this->json(['status' => 'error', 'message' => $e->getMessage()], Response::HTTP_BAD_REQUEST);
+        }
+
+        return $this->json(['status' => 'ok']);
+    }
+
     #[Route('/user/block', name: 'app_user_block', methods: ['POST'])]
     public function blockUser(Request $request): Response
     {
-        return $this->returnResponse(
-            $this->managementService->blockUser(
-                $request->request->all('userIds')
-            )
-        );
+        return $this->handleRequest($request, 'block');
     }
 
     #[Route('/user/unblock', name: 'app_user_unblock', methods: ['POST'])]
     public function unblockUser(Request $request): Response
     {
-        return $this->returnResponse(
-            $this->managementService->unblockUser(
-                $request->request->all('userIds')
-            )
-        );
+        return $this->handleRequest($request, 'unblock');
     }
 
     #[Route('/user/delete', name: 'app_user_delete', methods: ['POST'])]
     public function deleteUser(Request $request): Response
     {
-        return $this->returnResponse(
-            $this->managementService->deleteUser(
-                (array) $request->request->all('userIds'),
-                $request->getSession()
-            )
-        );
+        return $this->handleRequest($request, 'delete');
     }
 
     #[Route('/user/activate/{token}', name: 'app_user_activate', methods: ['GET'])]
     public function activateUser(string $token): Response
     {
-        if ($this->managementService->activateUser($token)) 
+        if ($this->managementService->activateUser($token))
             return $this->render('user/success_user_activation.html.twig');
 
         // i do not know what to do, so i just put it here
         $this->addFlash('error', 'Something went wrong!');
         return $this->redirectToRoute('app_index');
-    }
-
-    private function returnResponse(bool $result): Response
-    {
-        return $result
-            ? $this->json(['status' => 'ok'])
-            : $this->json(['status' => 'error'], Response::HTTP_BAD_REQUEST);
     }
 }
